0x0F-function_pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -7,7 +7,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-long unsigned int i;
+size_t i;
 if (array != 0)
 {
 for (i = 0; i < size; i++)
